Working out what to do with recipe multiplier 1
-Recipe multiplier 1x will add a shapeless recipe to convert the ore to
the vanilla ore. Only actually useful for ores that drop items, which is
vanilla only(? Need to check). Will be crafting not smelting as there
should be no penalty (i.e. time and fuel) if the multiplier is 1x.
Otherwise the 1x multiplier will simply add the ore to the dictionary.